+++ app/src/main/java/com/example/k_populare_libraries/UsersFragment.kt
@@ -4,7 +4,6 @@ import android.annotation.SuppressLint
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.k_populare_libraries.adapter.UsersRVAdapter
 import com.example.k_populare_libraries.app.App
@@ -13,7 +12,6 @@ import com.example.k_populare_libraries.presenter.UsersPresenter
 import com.example.k_populare_libraries.repository.GithubUsersRepo
 import com.example.k_populare_libraries.screens.AndroidScreens
 import com.example.k_populare_libraries.screens.BackButtonListenerI
-import com.example.k_populare_libraries.view.UserItemViewI
 import com.example.k_populare_libraries.view.UsersViewI
 import moxy.MvpAppCompatFragment
 import moxy.ktx.moxyPresenter
@@ -22,12 +20,12 @@ class UsersFragment : MvpAppCompatFragment(), UsersViewI, BackButtonListenerI {
 
     companion object {
         fun newInstance() = UsersFragment()
-
     }
 
     val presenter: UsersPresenter by moxyPresenter {
         UsersPresenter(GithubUsersRepo(), App.instance.router, AndroidScreens())
     }
+
     var adapter: UsersRVAdapter? = null
     private var binding: FragmentUsersBinding? = null
