+++ app/src/main/java/com/example/k_populare_libraries/adapter/UsersRVAdapter.kt
@@ -4,11 +4,13 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.k_populare_libraries.databinding.ItemRecyclerviewBinding
+import com.example.k_populare_libraries.presenter.UserPresenter
 import com.example.k_populare_libraries.view.UserItemViewI
 import com.example.k_populare_libraries.view.UserListPresenterI
 
 class UsersRVAdapter(val presenter: UserListPresenterI) :
     RecyclerView.Adapter<UsersRVAdapter.ViewHolder>() {
+
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) =
         ViewHolder(
@@ -17,14 +19,19 @@ class UsersRVAdapter(val presenter: UserListPresenterI) :
                 parent,
                 false
             )
-        ).apply {
-            itemView.setOnClickListener { presenter.itemClickListener?.invoke(this) }
-        }
+        )
 
     override fun getItemCount() = presenter.getCount()
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) =
-        presenter.bindView(holder.apply { pos = position })
+        presenter.bindView(holder.apply {
+            pos = position
+
+            itemView.setOnClickListener { view ->
+
+                presenter.itemClickListener?.invoke(this)
+            }
+        })
 
     inner class ViewHolder(val vb: ItemRecyclerviewBinding) : RecyclerView.ViewHolder(vb.root),
         UserItemViewI {
