+++ app/src/main/java/com/example/k_populare_libraries/presenter/UserPresenter.kt
@@ -1,0 +1,9 @@
+package com.example.k_populare_libraries.presenter
+
+import com.example.k_populare_libraries.view.UserViewI
+import moxy.MvpPresenter
+
+class UserPresenter(
+) :
+    MvpPresenter<UserViewI>() {
+}
