+++ app/src/main/java/com/example/k_populare_libraries/presenter/UsersPresenter.kt
@@ -1,29 +1,18 @@
 package com.example.k_populare_libraries.presenter
 
-import com.example.k_populare_libraries.data.GithubUser
 import com.example.k_populare_libraries.repository.GithubUsersRepo
 import com.example.k_populare_libraries.view.ScreensI
-import com.example.k_populare_libraries.view.UserItemViewI
-import com.example.k_populare_libraries.view.UserListPresenterI
 import com.example.k_populare_libraries.view.UsersViewI
 import com.github.terrakok.cicerone.Router
 import moxy.MvpPresenter
 
-class UsersPresenter(val userRepo: GithubUsersRepo, val router: Router, val screensI: ScreensI) :
+
+class UsersPresenter(
+    val userRepo: GithubUsersRepo,
+    val router: Router,
+    val screensI: ScreensI
+) :
     MvpPresenter<UsersViewI>() {
-
-    class UserListPresenter : UserListPresenterI {
-        val users = mutableListOf<GithubUser>()
-
-        override var itemClickListener: ((UserItemViewI) -> Unit)? = null
-
-        override fun bindView(view: UserItemViewI) {
-            val user = users[view.pos]
-            view.setLogin(user.login)
-        }
-
-        override fun getCount() = users.size
-    }
 
     val userListPresenter = UserListPresenter()
 
@@ -34,12 +23,10 @@ class UsersPresenter(val userRepo: GithubUsersRepo, val router: Router, val scre
 
         userListPresenter.itemClickListener = { itemView ->
 
-            router.setResultListener("STRING") {
-
-
-            }
+            //Log.d("TAG", userRepo.getUsers().get(itemView.pos).login)
             router.navigateTo(screensI.user())
         }
+
     }
 
     private fun loaData() {
