+++ app/src/main/java/com/example/k_populare_libraries/repository/GithubUsersRepo.kt
@@ -14,5 +14,4 @@ class GithubUsersRepo {
     fun getUsers() : List<GithubUser> {
         return repositories
     }
-
 }
