+++ app/src/main/java/com/example/k_populare_libraries/view/UserViewI.kt
@@ -1,0 +1,10 @@
+package com.example.k_populare_libraries.view
+
+import moxy.MvpView
+import moxy.viewstate.strategy.AddToEndSingleStrategy
+import moxy.viewstate.strategy.StateStrategyType
+
+@StateStrategyType(AddToEndSingleStrategy::class)
+interface UserViewI : MvpView {
+    fun set()
+}
